重构查询操作

+++ src/com/coderli/Dao/GroupDao.java
@@ -11,6 +11,6 @@ public interface GroupDao {
 	//根据学生的学号获取学生的签到信息
 	List<Sign> getSignInfo(String unumber, String page);
 	//根据学生的学号获取学生签到信息数量
-	int getSignCount(String unumber);
+	long getSignCount(String unumber);
 
 }
